fix: updates to support latest error codes and permissions
* changes expected AUTHENTICATION_ERRORs to PERMISSION_ERROR
* combines token-scope test with existing auth-client test, because
  token-scope test now requires session tokens
