+++ packages/client-sdk-nodejs/test/integration/shared/auth-client.test.ts
@@ -5,10 +5,14 @@ const {
   sessionTokenAuthClient,
   legacyTokenAuthClient,
   authTokenAuthClientFactory,
+  cacheClientFactory,
+  cacheName,
 } = SetupAuthClientIntegrationTest();
 
 runAuthClientTests(
   sessionTokenAuthClient,
   legacyTokenAuthClient,
-  authTokenAuthClientFactory
+  authTokenAuthClientFactory,
+  cacheClientFactory,
+  cacheName
 );
